+++ app/assets/stylesheets/gto/_nav.scss
@@ -89,7 +89,7 @@
       font-size: 0.75em;
     }
 
-    .nav-secondary-link-emphasis {
+    .nav-secondary-link--emphasis {
       font-size: 0.75em;
       color: $gto_orange;
     }
