+++ spec/models/artist_spec.rb
@@ -268,29 +268,6 @@ describe Artist do
     end
   end
 
-  describe 'remind_for_open_studios_register?' do
-    let!(:artist_to_remind) { create(:artist) }
-    let(:future_open_studios_event) { FactoryBot.create(:open_studios_event, :future) }
-
-    before do
-      Timecop.freeze
-      allow(OpenStudiosEventService).to receive(:current).and_return(future_open_studios_event)
-      artist.open_studios_events << future_open_studios_event
-    end
-
-    after do
-      Timecop.return
-    end
-
-    it 'returns true if artist has not registered for current event and it is no more than 12 weeks away' do
-      expect(artist_to_remind.remind_for_open_studios_register?).to eq(true)
-    end
-
-    it 'returns false if the artist is already registered for current event' do
-      expect(artist.remind_for_open_studios_register?).to eq(false)
-    end
-  end
-
   describe 'art piece helpers' do
     before do
       artist
